Add Unpack Methods
Uncompressing with tar and --xattrs is too much depending on the system
configuration which is being run on. For example, it seems with certain
versions of gnutar it's not possible to extract cap bits, and bsdtar should be
used instead.

With this change we glob methods used by moby, containerd and umoci to unpack layers
and thus removing the tar dependency when unpacking docker image layers.